requirements.txt
snowflake-ml-python>=1.5.5
snowflake-snowpark-python
pyyaml
